fix: 피드백 사항 반영

+++ src/main/java/kr/co/suitcarrier/web/dto/CartRequestDto.java
@@ -7,8 +7,8 @@ import lombok.Getter;
 @Getter
 public class CartRequestDto {
 
-    private Integer user;
-    private Integer post;
+    private Long user;
+    private Long post;
     private LocalDateTime rentDate;
     private LocalDateTime returnDate;
     private int rentPossible;
